fix(markers): run Intro & Credits like previews — respect workers, same queue, retries and recovery (#314)
* fix(markers): accuracy round 2 — audit follow-ups on sflix

- Credits: I Survived regressions fixed (the 640 read after an answer's end, a dense-text start); Accused 50/4, I Survived 14/1.
- One software shrink for the intro end-picture check too; GPU decode is checked once per card against the CPU (a diagnostic; GPU workers keep decoding on their GPU).
- Season audio: a file's speed is judged from its audio (RuPaul S08E04); an end-card rule (Tomb Raider King); guarded carry-over when a same-length replacement has no evidence.
- Decide rules: SkipDB needs a second source; credit text checks a credits chapter SkipDB disputes; online credits may overshoot the file end by 5 s; another release's intro beside season audio.
- A start-time version check re-runs files whose answers or decisions come from an older detector or older rules, in LOW batches ("Intro & Credits: re-checking files after an update").
- Older Plex answers dropped on kept-type items; a skipped detector is re-asked in the same run; the 2 s keep rule only with several versions.

Versions: credits text 5, season audio 9, end picture 3, decide rules 1.


* fix(markers): run Intro & Credits like previews — respect GPU/CPU workers, same queue, retries and recovery

Intro & Credits now follows the same rules as previews at every step:
- GPU workers keep their work on the GPU. Text detection retries the GPU with a
  back-off (5 s doubling to 10 min) instead of switching to the CPU because it
  is faster. The per-GPU decode check is diagnostic only.
- Per-worker ffmpeg_threads apply to marker decodes, as they do to previews;
  CPU text helpers scale with cpu_threads. Nothing is hard-coded.
- Pause freezes marker FFmpeg runs, and a paused state survives settings saves
  (settings schema v19).
- Busy files wait with a bound and hand back to the retry queue; scheduled and
  Season jobs follow the server pin; restarts revive waiting jobs by downtime.
- A cancel never stores "nothing found".

Previews:
- BIFs are written atomically.
- Partial runs are sorted by cause. Interrupted runs (SIGTERM/255, SIGKILL, the
  stall watchdog, I/O errors) are discarded, and the Dolby Vision fallback tiers
  are skipped. Crashes and self-exits keep the 95% publish rule and the DV-safe
  last resort.
- Exit 251 with GPU error output is treated as a GPU failure, not a disk error.
- Fallback messages name the real cause.

Also fixes three integration and GPU tests that already failed on the base
commit: the gate journey teardown, the fixture timeouts, and the vendor skip
guard.


---------

+++ media_preview_generator/processing/_embyish.py
@@ -34,7 +34,7 @@ class _EmbyishProcessor(_MediaServerProcessor):
         self,
         server_config: ServerConfig,
         *,
-        lookback_hours: int,
+        lookback_hours: float,
         library_ids: list[str] | None = None,
     ) -> Iterator[ProcessableItem]:
         """Walk items added in the last ``lookback_hours``.
