+++ media_preview_generator/processing/plex.py
@@ -36,7 +36,7 @@ class PlexProcessor(_MediaServerProcessor):
         self,
         server_config: ServerConfig,
         *,
-        lookback_hours: int,
+        lookback_hours: float,
         library_ids: list[str] | None = None,
     ) -> Iterator[ProcessableItem]:
         """Walk Plex sections for items added in the last ``lookback_hours``.
